+++ types/Country.d.ts
@@ -1,9 +1,0 @@
-import { Region } from './Region';
-
-export type Country = {
-  countryName: string;
-  phoneCode: string;
-  countryShortCode: string;
-  countryFlag: string;
-  regions: Region[];
-};